requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
cffi==1.16.0
cryptography==41.0.4
Django==4.2.5
django-bootstrap4==23.2
django-stdimage==6.0.1
gunicorn==21.2.0
mysql==0.0.3
mysql-client==0.0.1
mysqlclient==2.2.0
packaging==23.2
Pillow==10.0.1
pycparser==2.21
PyMySQL==1.1.0
soupsieve==2.5
sqlparse==0.4.4
whitenoise==6.5.0
